refactor(cli): distribute unit tests next to functions under test (#339)
Follow-up to #340: the unit-test module that previously lived at the
bottom of main.rs was hoisted into a single 1,714-LOC tests.rs during
the module split. This commit redistributes each test into the
production module that owns the function under test:

  - wrapper.rs        — stderr_indicates_unknown_session
  - zccache.rs        — rustc_wrapper_mode_from_env_var, is_sccache_wrapper
  - binaries.rs       — parse_tool_spec, rustup_resolution_failure,
                        known-tools registry sanity
  - cargo_front_door.rs — cargo arg parsing, low-disk warning
  - rust_plan.rs      — plan builder, thin manifest, warm restore,
                        bundle walk
  - gc.rs             — parse_gc_purge_answer, gc_purge_worker_count_for
  - main.rs           — Cli/Commands parsing, --as extraction,
                        self_relocate gate, version normalisation,
                        should_trampoline

Modules whose production code already exceeded 1,000 LOC before this
refactor (cargo_front_door, gc, rust_plan) link their test modules via
`#[cfg(test)] #[path = "*_tests.rs"] mod tests;` so the per-file LOC
ceiling stays as close to the original production count as possible.
main.rs uses the same sibling-file pattern to avoid tipping over the
ceiling. The rust_plan test suite is further split by topic
(plan_build, manifest, warm_restore, wire_compat, bundle_walk) so no
individual test file exceeds 1,000 LOC.

No behaviour changes; tests.rs is deleted.
