+++ crates/soldr-cli/src/cargo_front_door_tests.rs
@@ -1,0 +1,86 @@
+//! Unit tests for [`crate::cargo_front_door`]: the cargo argv parser, the
+//! low-disk warning helper, and the cargo-subcommand sniffer.
+//! Lives in a sibling file referenced via `#[path]` so `cargo_front_door.rs`
+//! stays comfortably under the 1000-LOC ceiling.
+
+use super::*;
+
+#[test]
+fn low_disk_warning_formats_yellow_below_threshold() {
+    let message = low_disk_warning_for_free_bytes(1536 * 1024 * 1024, true)
+        .expect("expected low-disk warning below threshold");
+    assert!(message.contains("\x1b[33mwarning\x1b[0m"));
+    assert!(message.contains("1.5 GB free"));
+    assert!(message.contains("Run `soldr gc`"));
+}
+
+#[test]
+fn low_disk_warning_omits_at_threshold() {
+    assert!(low_disk_warning_for_free_bytes(LOW_DISK_WARNING_THRESHOLD_BYTES, true).is_none());
+}
+
+#[test]
+fn low_disk_probe_failure_is_nonfatal() {
+    let warning = low_disk_warning_for_path(std::path::Path::new("."), true, |_| {
+        Err(std::io::Error::other("probe failed"))
+    });
+    assert!(warning.is_none());
+}
+
+#[test]
+fn cargo_args_detect_explicit_target_flag() {
+    assert!(cargo_args_specify_target(&[
+        "build".into(),
+        "--target".into(),
+        "x86_64-pc-windows-msvc".into(),
+    ]));
+    assert!(cargo_args_specify_target(&[
+        "build".into(),
+        "--target=x86_64-pc-windows-msvc".into(),
+    ]));
+}
+
+#[test]
+fn cargo_args_ignore_target_after_passthrough_separator() {
+    assert!(!cargo_args_specify_target(&[
+        "test".into(),
+        "--".into(),
+        "--target".into(),
+        "ignored".into(),
+    ]));
+}
+
+#[test]
+fn cargo_args_reject_reserved_no_cache_before_passthrough_separator() {
+    assert!(cargo_args_use_reserved_no_cache(&[
+        "build".into(),
+        "--no-cache".into(),
+    ]));
+    assert!(!cargo_args_use_reserved_no_cache(&[
+        "test".into(),
+        "--".into(),
+        "--no-cache".into(),
+    ]));
+}
+
+#[test]
+fn first_cargo_subcommand_skips_leading_flags() {
+    assert_eq!(
+        first_cargo_subcommand(&["--verbose".into(), "nextest".into(), "run".into()]),
+        Some("nextest")
+    );
+    assert_eq!(
+        first_cargo_subcommand(&["nextest".into(), "run".into()]),
+        Some("nextest")
+    );
+    assert_eq!(first_cargo_subcommand(&["--help".into()]), None);
+    assert_eq!(first_cargo_subcommand(&[]), None);
+}
+
+#[test]
+fn first_cargo_subcommand_stops_at_passthrough_separator() {
+    assert_eq!(
+        first_cargo_subcommand(&["--".into(), "nextest".into()]),
+        None
+    );
+}
